+++ src/client/shell/config.rs
@@ -133,6 +133,7 @@ impl ClientShellConfig {
             mouse_capture: config.ui.mouse_capture,
             mouse_scroll_lines: config.ui.mouse_scroll_lines(),
             right_click_passthrough_modifiers: config.ui.right_click_passthrough_modifiers(),
+            link_click_modifiers: config.ui.link_click_modifiers(),
             redraw_on_focus_gained: config.ui.redraw_on_focus_gained,
             switch_ascii_input_source_in_prefix: config
                 .experimental
@@ -323,6 +324,7 @@ impl ClientShellConfig {
                 self.mouse_capture = ui.mouse_capture;
                 self.mouse_scroll_lines = ui.mouse_scroll_lines();
                 self.right_click_passthrough_modifiers = ui.right_click_passthrough_modifiers();
+                self.link_click_modifiers = ui.link_click_modifiers();
                 self.redraw_on_focus_gained = ui.redraw_on_focus_gained;
             }
         }